requirements.txt
Flask>=3.0.0
chromadb>=0.5.0
python-dotenv>=1.0.1
PyMuPDF>=1.24.0
beautifulsoup4>=4.12.3
lxml>=5.2.0
gunicorn>=22.0.0
pytest>=8.0.0
